Split classes from Llama3
