+++ AOT.java
@@ -1,0 +1,58 @@
+package com.llama4j;
+
+import java.io.IOException;
+import java.nio.channels.FileChannel;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.StandardOpenOption;
+import java.util.Map;
+import java.util.Objects;
+
+final class AOT {
+    record PartialModel(String modelFileName, Llama model, long tensorDataOffset, Map<String, GGUF.GGUFTensorInfo> tensorInfos) {}
+
+    private static final PartialModel PRELOADED_GGUF = preLoadGGUF(System.getProperty("llama.PreloadGGUF"));
+
+    private static PartialModel preLoadGGUF(String modelPath) {
+        if (modelPath == null || modelPath.isEmpty()) {
+            return null;
+        }
+        try {
+            Path path = Path.of(modelPath);
+            if (!Files.exists(path) || !Files.isRegularFile(path)) {
+                throw new IllegalArgumentException("Cannot pre-load model: " + path);
+            }
+            GGUF gguf = GGUF.loadModel(path);
+            try (FileChannel fileChannel = FileChannel.open(path, StandardOpenOption.READ)) {
+                return new PartialModel(
+                        path.getFileName().toString(),
+                        ModelLoader.loadModel(fileChannel, gguf, Options.DEFAULT_MAX_TOKENS, false),
+                        gguf.getTensorDataOffset(),
+                        gguf.getTensorInfos()
+                );
+            }
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    public static Llama tryUsePreLoaded(Path modelPath, int contextLength) throws IOException {
+        AOT.PartialModel preLoaded = AOT.PRELOADED_GGUF;
+        if (preLoaded == null) {
+            return null;
+        }
+        String optionsModel = modelPath.getFileName().toString();
+        String preLoadedModel = preLoaded.modelFileName();
+        if (!Objects.equals(optionsModel, preLoadedModel)) {
+            return null;
+        }
+        Llama baseModel = preLoaded.model();
+        try (var timer = Timer.log("Load tensors from pre-loaded model");
+             var fileChannel = FileChannel.open(modelPath, StandardOpenOption.READ)) {
+            Map<String, GGMLTensorEntry> tensorEntries = GGUF.loadTensors(fileChannel, preLoaded.tensorDataOffset(), preLoaded.tensorInfos());
+            Llama.Weights weights = ModelLoader.loadWeights(tensorEntries, baseModel.configuration());
+            return new Llama(baseModel.configuration().withContextLength(contextLength), baseModel.tokenizer(), weights);
+        }
+    }
+}
+
