+++ ChatFormat.java
@@ -1,0 +1,81 @@
+package com.llama4j;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
+
+class ChatFormat {
+
+    final Tokenizer tokenizer;
+    final int beginOfText;
+    final int endHeader;
+    final int startHeader;
+    final int endOfTurn;
+    final int endOfText;
+    final int endOfMessage;
+    final Set<Integer> stopTokens;
+
+    public ChatFormat(Tokenizer tokenizer) {
+        this.tokenizer = tokenizer;
+        Map<String, Integer> specialTokens = this.tokenizer.getSpecialTokens();
+        this.beginOfText = specialTokens.get("<|begin_of_text|>");
+        this.startHeader = specialTokens.get("<|start_header_id|>");
+        this.endHeader = specialTokens.get("<|end_header_id|>");
+        this.endOfTurn = specialTokens.get("<|eot_id|>");
+        this.endOfText = specialTokens.get("<|end_of_text|>");
+        this.endOfMessage = specialTokens.getOrDefault("<|eom_id|>", -1);
+        this.stopTokens = Set.of(endOfText, endOfTurn);
+    }
+
+    public Tokenizer getTokenizer() {
+        return tokenizer;
+    }
+
+    public Set<Integer> getStopTokens() {
+        return stopTokens;
+    }
+
+    public List<Integer> encodeHeader(ChatFormat.Message message) {
+        List<Integer> tokens = new ArrayList<>();
+        tokens.add(startHeader);
+        tokens.addAll(this.tokenizer.encodeAsList(message.role().name()));
+        tokens.add(endHeader);
+        tokens.addAll(this.tokenizer.encodeAsList("\n"));
+        return tokens;
+    }
+
+    public List<Integer> encodeMessage(ChatFormat.Message message) {
+        List<Integer> tokens = this.encodeHeader(message);
+        tokens.addAll(this.tokenizer.encodeAsList(message.content().strip()));
+        tokens.add(endOfTurn);
+        return tokens;
+    }
+
+    public List<Integer> encodeDialogPrompt(boolean appendAssistantTurn, List<ChatFormat.Message> dialog) {
+        List<Integer> tokens = new ArrayList<>();
+        tokens.add(beginOfText);
+        for (ChatFormat.Message message : dialog) {
+            tokens.addAll(this.encodeMessage(message));
+        }
+        if (appendAssistantTurn) {
+            tokens.addAll(this.encodeHeader(new ChatFormat.Message(ChatFormat.Role.ASSISTANT, "")));
+        }
+        return tokens;
+    }
+
+    public record Message(ChatFormat.Role role, String content) {
+    }
+
+    public record Role(String name) {
+        public static ChatFormat.Role SYSTEM = new ChatFormat.Role("system");
+        public static ChatFormat.Role USER = new ChatFormat.Role("user");
+        public static ChatFormat.Role ASSISTANT = new ChatFormat.Role("assistant");
+
+        @Override
+        public String toString() {
+            return name;
+        }
+    }
+}
+
