+++ GGMLTensorEntry.java
@@ -1,0 +1,8 @@
+package com.llama4j;
+
+import java.lang.foreign.MemorySegment;
+
+public record GGMLTensorEntry(MemorySegment mappedFile, String name, GGMLType ggmlType, int[] shape,
+                       MemorySegment memorySegment) {
+}
+
