+++ Llama3.java
@@ -188,1730 +188,3 @@ public class Llama3 {
 
 
 
-record Llama(Configuration configuration, Tokenizer tokenizer, Weights weights) {
-    public State createNewState(int batchsize) {
-        State state = new State(configuration(), batchsize);
-        state.latestToken = tokenizer.getSpecialTokens().get("<|begin_of_text|>");
-        return state;
-    }
-
-    public static final class Configuration {
-        public final int dim; // transformer dimension
-        public final int hiddenDim; // for ffn layers
-        public final int numberOfLayers; // number of layers
-        public final int numberOfHeads; // number of query heads
-        public final int numberOfKeyValueHeads; // number of key/value heads (can be < query heads because of multiquery)
-        public final int vocabularySize; // vocabulary size, usually 256 (byte-level)
-        public final int contextLength; // max sequence length
-        public final float rmsNormEps;
-        public final float ropeTheta;
-        public final int headSize;
-
-        Configuration withContextLength(int newContextLength) {
-            if (newContextLength < 0) {
-                return this; // no change
-            }
-            return new Configuration(this.dim, this.hiddenDim, this.numberOfLayers, this.numberOfHeads, this.numberOfKeyValueHeads, this.vocabularySize, newContextLength, this.rmsNormEps, this.ropeTheta);
-        }
-
-        public Configuration(int dim, int hiddenDim, int numberOfLayers, int numberOfHeads, int numberOfKeyValueHeads, int vocabularySize, int contextLength, float rmsNormEps, float ropeTheta) {
-            this.dim = dim;
-            this.hiddenDim = hiddenDim;
-            this.numberOfLayers = numberOfLayers;
-            this.numberOfHeads = numberOfHeads;
-            this.numberOfKeyValueHeads = numberOfKeyValueHeads;
-            this.vocabularySize = vocabularySize;
-            this.contextLength = contextLength;
-            this.rmsNormEps = rmsNormEps;
-            this.ropeTheta = ropeTheta;
-            this.headSize = dim / numberOfHeads;
-        }
-    }
-
-    public static final class Weights {
-        // token embedding table
-        public final FloatTensor token_embedding_table; // (vocab_size, dim)
-        // weights for rmsnorms
-        public final FloatBuffer[] rms_att_weight; // (layer, dim) rmsnorm weights
-        // weights for matmuls
-        public final FloatTensor[] wq; // (layer, n_heads * head_size)
-        public final FloatTensor[] wk; // (layer, n_kv_heads, head_size)
-        public final FloatTensor[] wv; // (layer, n_kv_heads * head_size)
-        public final FloatTensor[] wo; // (layer, n_heads * head_size, dim)
-        public final FloatBuffer[] rms_ffn_weight; // (layer, dim)
-        // weights for ffn
-        public final FloatTensor[] w1; // (layer, hidden_dim, dim)
-        public final FloatTensor[] w2; // (layer, dim, hidden_dim)
-        public final FloatTensor[] w3; // (layer, hidden_dim, dim)
-        // public final rmsnorm
-        public final FloatBuffer rms_final_weight; // (dim,)
-        // freq_cis for RoPE relatively positional embeddings
-        public final FloatBuffer freq_cis_real; // (seq_len, head_size/2)
-        public final FloatBuffer freq_cis_imag; // (seq_len, head_size/2)
-        // (optional) classifier weights for the logits, on the last layer
-        public final FloatTensor wcls; // (vocab_size, dim)
-
-        public Weights(FloatTensor token_embedding_table, FloatBuffer[] rms_att_weight, FloatTensor[] wq, FloatTensor[] wk, FloatTensor[] wv, FloatTensor[] wo, FloatBuffer[] rms_ffn_weight, FloatTensor[] w1, FloatTensor[] w2, FloatTensor[] w3, FloatBuffer rms_final_weight, FloatBuffer freq_cis_real, FloatBuffer freq_cis_imag, FloatTensor wcls) {
-            this.token_embedding_table = token_embedding_table;
-            this.rms_att_weight = rms_att_weight;
-            this.wq = wq;
-            this.wk = wk;
-            this.wv = wv;
-            this.wo = wo;
-            this.rms_ffn_weight = rms_ffn_weight;
-            this.w1 = w1;
-            this.w2 = w2;
-            this.w3 = w3;
-            this.rms_final_weight = rms_final_weight;
-            this.freq_cis_real = freq_cis_real;
-            this.freq_cis_imag = freq_cis_imag;
-            this.wcls = wcls;
-        }
-    }
-
-    public static final class State {
-
-        // current wave of activations
-        public final int batchsize;
-        public final FloatTensor[] x; // activation at current time stamp (dim,)
-        public final FloatTensor[] xb; // same, but inside a residual branch (dim,)
-        public final FloatTensor[] xb2; // an additional buffer just for convenience (dim,)
-        public final FloatTensor[] hb; // buffer for hidden dimension in the ffn (hidden_dim,)
-        public final FloatTensor[] hb2; // buffer for hidden dimension in the ffn (hidden_dim,)
-        public final FloatTensor[] q; // query (dim,)
-        public final FloatTensor[] k; // key (dim,)
-        public final FloatTensor[] v; // value (dim,)
-        public final FloatTensor[] att; // buffer for scores/attention values (n_heads, seq_len)
-        public final FloatTensor logits; // output logits
-
-        // kv cache
-        public final FloatTensor[] keyCache;   // (n_layer, seq_len, kv_dim)
-        public final FloatTensor[] valueCache; // (n_layer, seq_len, kv_dim)
-        
-        /** last index in previous block */
-        int idxPrevBlock;
-
-        public int latestToken;
-
-        State(Configuration config, int batchsize) {
-            this.batchsize = batchsize;
-            this.x = allocate(batchsize, config.dim);
-            this.xb = allocate(batchsize, config.dim);
-            this.xb2 = allocate(batchsize, config.dim);
-            this.hb = allocate(batchsize, config.hiddenDim);
-            this.hb2 = allocate(batchsize, config.hiddenDim);
-            this.q = allocate(batchsize, config.dim);
-            this.k = allocate(batchsize, config.dim);
-            this.v = allocate(batchsize, config.dim);
-            this.att = allocate(batchsize, config.numberOfHeads, config.contextLength);
-            idxPrevBlock = -1;
-
-            this.logits = ArrayFloatTensor.allocate(config.vocabularySize);
-            int kvDim = (config.dim * config.numberOfKeyValueHeads) / config.numberOfHeads;
-            this.keyCache = Stream.generate(() -> ArrayFloatTensor.allocate(config.contextLength, kvDim)).limit(config.numberOfLayers).toArray(FloatTensor[]::new);
-            this.valueCache = Stream.generate(() -> ArrayFloatTensor.allocate(config.contextLength, kvDim)).limit(config.numberOfLayers).toArray(FloatTensor[]::new);
-        }
-    }
-
-    static FloatTensor[] allocate(int numTokens, int... dims) {
-        return IntStream.range(0, numTokens)
-                .mapToObj(i -> ArrayFloatTensor.allocate(dims))
-                .toArray(FloatTensor[]::new);
-    }
-
-    static void rmsnorm(FloatTensor out, FloatTensor x, FloatBuffer weight, int size, float rmsNormEps) {
-        // calculate sum of squares
-        float ss = x.reduce(0, size, 0f, (acc, xi) -> acc + xi * xi);
-        ss /= size;
-        ss += rmsNormEps;
-        ss = (float) (1.0 / Math.sqrt(ss));
-        // normalize and scale
-        final float finalss = ss; // for the lambda
-        out.mapWithIndexInPlace(0, size, (value, index) -> weight.get(index) * (finalss * x.getFloat(index)));
-    }
-
-    static FloatTensor forward(Llama model, State state, int[] tokens, int position, boolean computeLogits) {
-        // a few convenience variables
-        Configuration config = model.configuration();
-        Weights weights = model.weights();
-        int dim = config.dim;
-        int headSize = config.headSize;
-        int kvDim = (config.dim * config.numberOfKeyValueHeads) / config.numberOfHeads;
-        int kvMul = config.numberOfHeads / config.numberOfKeyValueHeads; // integer multiplier of the kv sharing in multiquery
-        float sqrtHeadSize = (float) Math.sqrt(headSize);
-        final int nTokens = tokens.length;
-
-        // copy the token embedding into x
-        Parallel.parallelFor(0, nTokens, t ->
-            weights.token_embedding_table.copyTo(tokens[t] * dim, state.x[t], 0, dim)
-        );
-
-        // forward all the layers
-        for (int l = 0; l < config.numberOfLayers; l++) {
-            // attention rmsnorm
-            // rmsnorm(state.xb, state.x, weights.rms_att_weight[l], dim, config.rmsNormEps);
-            final int curLayer = l;
-            Parallel.parallelFor(0, nTokens, t ->
-                rmsnorm(state.xb[t], state.x[t], weights.rms_att_weight[curLayer], dim, config.rmsNormEps)
-            );
-
-            // qkv matmuls for this position
-            weights.wq[l].matmul(nTokens, state.xb, state.q, dim, dim);
-            weights.wk[l].matmul(nTokens, state.xb, state.k, kvDim, dim);
-            weights.wv[l].matmul(nTokens, state.xb, state.v, kvDim, dim);
-
-            // RoPE relative positional encoding: complex-valued rotate q and k in each head
-            Parallel.parallelFor(0, nTokens, t -> {
-                for (int i = 0; i < dim; i += 2) {
-                    int head_dim = i % headSize;
-                    float fcr = weights.freq_cis_real.get((position + t) * (headSize / 2) + (head_dim / 2));
-                    float fci = weights.freq_cis_imag.get((position + t) * (headSize / 2) + (head_dim / 2));
-                    int rotn = i < kvDim ? 2 : 1; // how many vectors? 2 = q & k, 1 = q only
-                    for (int vi = 0; vi < rotn; vi++) {
-                        FloatTensor vec = vi == 0 ? state.q[t] : state.k[t]; // the vector to rotate (query or key)
-                        float v0 = vec.getFloat(i);
-                        float v1 = vec.getFloat(i + 1);
-                        vec.setFloat(i, v0 * fcr - v1 * fci);
-                        vec.setFloat(i + 1, v0 * fci + v1 * fcr);
-                    }
-                }
-            });
-
-            // save key,value at this time step (position) to our kv cache
-            //int loff = l * config.seq_len * kvDim; // kv cache layer offset for convenience
-            Parallel.parallelFor(0, nTokens, t -> {
-                state.k[t].copyTo(0, state.keyCache[curLayer], (position + t) * kvDim, kvDim);
-                state.v[t].copyTo(0, state.valueCache[curLayer], (position + t) * kvDim, kvDim);
-            });
-
-            // If the logits are not required, the attention and FFN of the last layer can be skipped entirely.
-            if (!computeLogits && curLayer == config.numberOfLayers - 1) {
-                state.idxPrevBlock = nTokens - 1;
-                return null;
-            }
-
-            // multihead attention. iterate over all heads
-            Parallel.parallelForLong(0, (long) nTokens * (long) config.numberOfHeads, ht -> {
-                int token = (int) (ht / config.numberOfHeads);
-                int h = (int) (ht % config.numberOfHeads);
-                // get the query vector for this head
-                // float* q = s.q + h * headSize;
-                int qOffset = h * headSize;
-
-                // attention scores for this head
-                // float* att = s.att + h * config.seq_len;
-                int attOffset = h * config.contextLength;
-
-                // iterate over all timesteps, including the current one
-                for (int t = 0; t <= position + token; t++) {
-                    // get the key vector for this head and at this timestep
-                    // float* k = s.key_cache + loff + t * dim + h * headSize;
-                    int keyCacheOffset = /* loff + */ t * kvDim + (h / kvMul) * headSize;
-                    // calculate the attention score as the dot product of q and k
-                    float score = state.q[token].dot(qOffset, state.keyCache[curLayer], keyCacheOffset, headSize);
-                    score /= sqrtHeadSize;
-                    // save the score to the attention buffer
-                    state.att[token].setFloat(attOffset + t, score);
-                }
-
-                // softmax the scores to get attention weights, from 0..position inclusively
-                state.att[token].softmaxInPlace(attOffset, position + token + 1);
-
-                // weighted sum of the values, store back into xb
-                // float* xb = s.xb + h * headSize;
-                int xbOffset = h * headSize;
-                // memset(xb, 0, headSize * sizeof(float));
-                state.xb[token].fillInPlace(xbOffset, headSize, 0f);
-
-                for (int t = 0; t <= position + token; t++) {
-                    // get the value vector for this head and at this timestep
-                    // float* v = s.value_cache + loff + t * dim + h * headSize;
-                    int vOffset = /* loff + */ t * kvDim + (h / kvMul) * headSize;
-                    // get the attention weight for this timestep
-                    float a = state.att[token].getFloat(attOffset + t);
-                    // accumulate the weighted value into xb
-                    state.xb[token].saxpyInPlace(xbOffset, state.valueCache[curLayer], vOffset, headSize, a);
-                }
-            });
-
-            // final matmul to get the output of the attention
-            weights.wo[l].matmul(nTokens, state.xb, state.xb2, dim, dim);
-
-            // residual connection back into x
-            Parallel.parallelFor(0, nTokens, t -> {
-                state.x[t].addInPlace(state.xb2[t]);
-            });
-
-            // ffn rmsnorm
-            Parallel.parallelFor(0, nTokens, t -> {
-                rmsnorm(state.xb[t], state.x[t], weights.rms_ffn_weight[curLayer], dim, config.rmsNormEps);
-            });
-
-            // Now for FFN in PyTorch we have: self.w2(F.silu(self.w1(x)) * self.w3(x))
-            // first calculate self.w1(x) and self.w3(x)
-            weights.w1[l].matmul(nTokens, state.xb, state.hb, config.hiddenDim, dim);
-            weights.w3[l].matmul(nTokens, state.xb, state.hb2, config.hiddenDim, dim);
-
-            // SwiGLU non-linearity
-            // silu(x)=x*σ(x), where σ(x) is the logistic sigmoid
-            Parallel.parallelFor(0, nTokens, t -> {
-                state.hb[t].mapInPlace(value -> value / (float) (1.0 + Math.exp(-value)));
-            });
-
-            // elementwise multiply with w3(x)
-            Parallel.parallelFor(0, nTokens, t -> {
-                state.hb[t].multiplyInPlace(state.hb2[t]);
-            });
-
-            // final matmul to get the output of the ffn
-            weights.w2[l].matmul(nTokens, state.hb, state.xb, dim, config.hiddenDim);
-
-            // residual connection
-            Parallel.parallelFor(0, nTokens, t -> {
-                state.x[t].addInPlace(state.xb[t]);
-            });
-        }
-
-        // final rmsnorm
-        Parallel.parallelFor(0, nTokens, t -> {
-            rmsnorm(state.x[t], state.x[t], weights.rms_final_weight, dim, config.rmsNormEps);
-        });
-
-        // classifier into logits
-        weights.wcls.matmul(state.x[nTokens - 1], state.logits, config.vocabularySize, dim);
-        state.idxPrevBlock = nTokens - 1;
-
-        return state.logits;
-    }
-
-    /**
-     * LLM generation entry point, ingest prompt tokens and generates new tokens.
-     *
-     * <p>
-     * All prompt tokens are ingested first, then inference starts, until a stop token is found.
-     * The returned tokens only include generated/inferred tokens.
-     *
-     * @param model            model to run inference (including weights, configuration, tokenizer ...)
-     * @param state            state of the model e.g. key/value caches ... this is mutated by this call
-     * @param startPosition    start prompt ingestion + inference at this position in the context e.g. useful if state was kept across calls (chained generation). 0 implies run with no previous context.
-     * @param promptTokens     prompt tokens to ingest, all the prompt tokens will be ingested, given there's enough capacity left in the context
-     * @param stopTokens       set of tokens that abort generation during inference, stop tokens do not affect prompt ingestion
-     * @param maxTokens        maximum number of tokens (can go up to {@link Configuration#contextLength context length}
-     *                         if this value is negative or greater than {@link Configuration#contextLength context length}
-     * @param sampler          {@link Sampler strategy} used to select tokens
-     * @param echo             debugging flag, prints ALL, prompt and inferred tokens, to {@link System#err stderr}
-     * @param onTokenGenerated callback, if non-null, it's called every time a token is inferred e.g. it's not called when ingesting prompt tokens
-     * @return list of generated/inferred tokens, including the stop token, if any e.g. does not include any token from the prompt
-     */
-    public static List<Integer> generateTokens(Llama model, State state, int startPosition, List<Integer> promptTokens, Set<Integer> stopTokens, int maxTokens, Sampler sampler, boolean echo,
-                                               IntConsumer onTokenGenerated) {
-        long startNanos = System.nanoTime();
-        long startGen = 0;
-        if (maxTokens < 0 || model.configuration().contextLength < maxTokens) {
-            maxTokens = model.configuration().contextLength;
-        }
-        List<Integer> generatedTokens = new ArrayList<>(maxTokens);
-        int token = state.latestToken; // BOS?
-        int nextToken;
-        int promptIndex = 0;
-        for (int position = startPosition; position < maxTokens; ++position) {
-            if (promptIndex < promptTokens.size()) {
-                final int nTokens = Math.min(maxTokens - position, Math.min(promptTokens.size() - promptIndex, state.batchsize));
-                final int[] tokens = new int[nTokens];
-                for (int i = 0; i < nTokens; i++) {
-                    tokens[i] = promptTokens.get(promptIndex + i);
-                    if (echo) {
-                        // log prompt token (different color?)
-                        System.err.print(Tokenizer.replaceControlCharacters(model.tokenizer().decode(List.of(tokens[i]))));
-                    }
-                }
-                if (echo) {
-                    System.out.format("position=%d, promptIdx=%d, promptSize=%d, tokens=%s%n", position, promptIndex, promptTokens.size(), Arrays.toString(tokens));
-                }
-                // Only compute logits on the very last batch.
-                boolean computeLogits = promptIndex + nTokens >= promptTokens.size();
-                forward(model, state, tokens, position, computeLogits);
-                position += nTokens - 1; // -1 -> incremented later in the for loop
-                promptIndex += nTokens;
-                if (promptIndex < promptTokens.size()) {
-                    continue;
-                }
-                startGen = System.nanoTime();
-            } else {
-                forward(model, state, new int[]{token}, position, true);
-            }
-            nextToken = sampler.sampleToken(state.logits);
-            if (echo) {
-                // log inferred token
-                System.err.print(Tokenizer.replaceControlCharacters(model.tokenizer().decode(List.of(nextToken))));
-            }
-            generatedTokens.add(nextToken);
-            if (onTokenGenerated != null) {
-                onTokenGenerated.accept(nextToken);
-            }
-            if (stopTokens.contains(nextToken)) {
-                break;
-            }
-            state.latestToken = token = nextToken;
-        }
-
-        long elapsedNanos = System.nanoTime() - startNanos;
-        long promptNanos = startGen - startNanos;
-        long genNanos = elapsedNanos - startGen + startNanos;
-        System.err.printf("%ncontext: %d/%d prompt: %.2f tokens/s (%d) generation: %.2f tokens/s (%d)%n",
-                startPosition + promptIndex + generatedTokens.size(), model.configuration().contextLength,
-                promptTokens.size() / (promptNanos / 1_000_000_000.0), promptTokens.size(),
-                generatedTokens.size() / (genNanos / 1_000_000_000.0), generatedTokens.size());
-
-        return generatedTokens;
-    }
-}
-
-/**
- * Byte Pair Encoding tokenizer.
- * <p>
- * Based on <a href="https://github.com/karpathy/minbpe">minbpe</a>, algorithmically follows along the
- * <a href="https://github.com/openai/gpt-2/blob/master/src/encoder.py">GPT 2 tokenizer</a>
- */
-class Tokenizer {
-    private final Pattern compiledPattern;
-    private final Vocabulary vocabulary;
-    private final Map<Pair<Integer, Integer>, Integer> merges;
-    private final Map<String, Integer> specialTokens;
-
-    public String regexPattern() {
-        if (compiledPattern == null) {
-            return null;
-        }
-        return compiledPattern.pattern();
-    }
-
-    public Map<String, Integer> getSpecialTokens() {
-        return specialTokens;
-    }
-
-    public boolean isSpecialToken(int tokenIndex) {
-        return specialTokens.containsValue(tokenIndex);
-    }
-
-    public Tokenizer(Vocabulary vocabulary, List<Pair<Integer, Integer>> merges, String regexPattern, Map<String, Integer> specialTokens) {
-        this.vocabulary = vocabulary;
-        this.compiledPattern = regexPattern != null ? Pattern.compile(regexPattern) : null;
-        this.specialTokens = new HashMap<>(specialTokens);
-        this.merges = new HashMap<>();
-        for (Pair<Integer, Integer> pair : merges) {
-            int firstIndex = pair.first();
-            int secondIndex = pair.second();
-            int mergeIndex = vocabulary.getIndex(vocabulary.get(firstIndex) + vocabulary.get(secondIndex)).orElseThrow();
-            this.merges.put(pair, mergeIndex);
-        }
-    }
-
-    private int[] encodeImpl(String text) {
-        return encode(text, Set.of()).stream().mapToInt(i -> i).toArray();
-    }
-
-    /**
-     * Unlike {@link #encodeOrdinary(String)}, this function handles special tokens.
-     * allowed_special: can be "all"|"none"|"none_raise" or a custom set of special tokens
-     * if none_raise, then an error is raised if any special token is encountered in text
-     * this is the default tiktoken behavior right now as well
-     * any other behavior is either annoying, or a major footgun.
-     */
-    List<Integer> encode(String text, Set<String> allowedSpecial) {
-        // decode the user desire w.r.t. handling of special tokens
-        Set<String> special = allowedSpecial;
-        assert getSpecialTokens().keySet().containsAll(special);
-        if (special.isEmpty()) {
-            // shortcut: if no special tokens, just use the ordinary encoding
-            return encodeOrdinary(text);
-        }
-
-        // otherwise, we have to be careful with potential special tokens in text
-        // we handle special tokens by splitting the text
-        // based on the occurrence of any exact match with any of the special tokens
-        // we can use re.split for this. note that surrounding the pattern with ()
-        // makes it into a capturing group, so the special tokens will be included
-        String specialPattern = special
-                .stream()
-                .map(Pattern::quote)
-                .collect(Collectors.joining("|", "(", ")"));
-
-        String[] specialChunks = text.split(specialPattern);
-        // now all the special characters are separated from the rest of the text
-        // all chunks of text are encoded separately, then results are joined
-        List<Integer> ids = new ArrayList<>();
-        for (String part : specialChunks) {
-            if (special.contains(part)) {
-                // this is a special token, encode it separately as a special case
-                ids.add(getSpecialTokens().get(part));
-            } else {
-                // this is an ordinary sequence, encode it normally
-                ids.addAll(encodeOrdinary(part));
-            }
-        }
-        return ids;
-    }
-
-    private static List<String> findAll(Pattern pattern, String text) {
-        List<String> allMatches = new ArrayList<>();
-        Matcher matcher = pattern.matcher(text);
-        while (matcher.find()) {
-            allMatches.add(matcher.group());
-        }
-        return allMatches;
-    }
-
-    /**
-     * Encoding that ignores any special tokens.
-     */
-    public List<Integer> encodeOrdinary(String text) {
-        // split text into chunks of text by categories defined in regex pattern
-        List<String> textChunks = findAll(compiledPattern, text);
-        // all chunks of text are encoded separately, then results are joined
-        List<Integer> ids = new ArrayList<>();
-        for (String chunk : textChunks) {
-            List<Integer> chunkIds = encodeChunk(chunk);
-            ids.addAll(chunkIds);
-        }
-        return ids;
-    }
-
-    private Map<Pair<Integer, Integer>, Integer> getStats(List<Integer> ids) {
-        Map<Pair<Integer, Integer>, Integer> map = new HashMap<>();
-        for (int i = 0; i + 1 < ids.size(); i++) {
-            Pair<Integer, Integer> key = new Pair<>(ids.get(i), ids.get(i + 1));
-            map.put(key, map.getOrDefault(key, 0) + 1);
-        }
-        return map;
-    }
-
-    private List<Integer> encodeChunk(String chunk) {
-        // return the token ids
-        // let's begin. first, convert all bytes to integers in range 0..255
-        List<Integer> ids = new ArrayList<>();
-        for (int b : chunk.toCharArray()) {
-            int tokenIndex = this.vocabulary.getIndex(String.valueOf((char) b)).orElseThrow();
-            ids.add(tokenIndex);
-        }
-
-        while (ids.size() >= 2) {
-            // find the pair with the lowest merge index
-            Map<Pair<Integer, Integer>, Integer> stats = getStats(ids);
-            Pair<Integer, Integer> pair = stats.keySet().stream().min(Comparator.comparingInt(key -> this.merges.getOrDefault(key, Integer.MAX_VALUE))).orElseThrow();
-            // subtle: if there are no more merges available, the key will
-            // result in an inf for every single pair, and the min will be
-            // just the first pair in the list, arbitrarily
-            // we can detect this terminating case by a membership check
-            if (!this.merges.containsKey(pair)) {
-                break; // nothing else can be merged anymore
-            }
-            // otherwise let's merge the best pair (lowest merge index)
-            int idx = this.merges.get(pair);
-            ids = merge(ids, pair, idx);
-        }
-        return ids;
-    }
-
-    private static List<Integer> merge(List<Integer> ids, Pair<Integer, Integer> pair, int idx) {
-        List<Integer> newids = new ArrayList<>();
-        int i = 0;
-        while (i < ids.size()) {
-            // if not at the very last position AND the pair matches, replace it
-            if (ids.get(i).equals(pair.first()) && i < ids.size() - 1 && ids.get(i + 1).equals(pair.second())) {
-                newids.add(idx);
-                i += 2;
-            } else {
-                newids.add(ids.get(i));
-                i += 1;
-            }
-        }
-        return newids;
-    }
-
-    public String decodeImpl(List<Integer> tokens) {
-        StringBuilder sb = new StringBuilder();
-        for (int token : tokens) {
-            String tokenString = vocabulary.get(token);
-            sb.append(tokenString);
-        }
-        return sb.toString();
-    }
-
-    /**
-     * Returns list of utf-8 byte and a corresponding list of unicode strings.
-     * The reversible bpe codes work on unicode strings.
-     * This means you need a large # of unicode characters in your vocab if you want to avoid UNKs.
-     * When you're at something like a 10B token dataset you end up needing around 5K for decent coverage.
-     * This is a significant percentage of your normal, say, 32K bpe vocab.
-     * To avoid that, we want lookup tables between utf-8 bytes and unicode strings.
-     * And avoids mapping to whitespace/control characters the bpe code barfs on.
-     */
-    private static Map<Integer, Integer> bytesToUnicode() {
-        List<Integer> bs = new ArrayList<>();
-        IntStream.rangeClosed('!', '~').forEach(bs::add);
-        IntStream.rangeClosed('¡', '¬').forEach(bs::add);
-        IntStream.rangeClosed('®', 'ÿ').forEach(bs::add);
-
-        List<Integer> cs = new ArrayList<>(bs);
-        int n = 0;
-        for (int b = 0; b < 256; ++b) {
-            if (!bs.contains(b)) {
-                bs.add(b);
-                cs.add(256 + n);
-                n += 1;
-            }
-        }
-
-        // return dict(zip(bs, cs))
-        return IntStream.range(0, bs.size())
-                .boxed()
-                .collect(Collectors.toMap(bs::get, cs::get));
-    }
-
-    static final Map<Integer, Integer> BYTE_ENCODER = bytesToUnicode();
-    static final Map<Integer, Integer> BYTE_DECODER = BYTE_ENCODER.entrySet()
-            .stream()
-            .collect(Collectors.toMap(Map.Entry::getValue, Map.Entry::getKey));
-
-    public int[] encode(String text) {
-        StringBuilder sb = new StringBuilder();
-        byte[] bytes = text.getBytes(StandardCharsets.UTF_8);
-        for (byte b : bytes) {
-            sb.appendCodePoint(BYTE_ENCODER.get(Byte.toUnsignedInt(b)));
-        }
-        return encodeImpl(sb.toString());
-    }
-
-    public static String replaceControlCharacters(int[] codePoints) {
-        // we don't want to print control characters
-        // which distort the output (e.g. \n or much worse)
-        // https://stackoverflow.com/questions/4324790/removing-control-characters-from-a-string-in-python/19016117#19016117
-        // http://www.unicode.org/reports/tr44/#GC_Values_Table\
-        StringBuilder chars = new StringBuilder();
-        for (int cp : codePoints) {
-            if (Character.getType(cp) == Character.CONTROL && cp != '\n') {
-                chars.append("\\u").append(HexFormat.of().toHexDigits(cp, 4)); // escape
-            } else {
-                chars.appendCodePoint(cp); // this character is ok
-            }
-        }
-        return chars.toString();
-    }
-
-    public static String replaceControlCharacters(String str) {
-        return replaceControlCharacters(str.codePoints().toArray());
-    }
-
-    public List<Integer> encodeAsList(String text) {
-        return Arrays.stream(encode(text)).boxed().toList();
-    }
-
-    public String decode(List<Integer> tokens) {
-        String decoded = decodeImpl(tokens);
-        int[] decodedBytesAsInts = decoded.codePoints().map(BYTE_DECODER::get).toArray();
-        byte[] rawBytes = new byte[decodedBytesAsInts.length];
-        for (int i = 0; i < decoded.length(); i++) {
-            rawBytes[i] = (byte) decodedBytesAsInts[i];
-        }
-        return new String(rawBytes, StandardCharsets.UTF_8);
-    }
-}
-
-final class Parallel {
-    public static void parallelFor(int startInclusive, int endExclusive, IntConsumer action) {
-        if (startInclusive == 0 && endExclusive == 1) {
-            action.accept(0);
-            return;
-        }
-        IntStream.range(startInclusive, endExclusive).parallel().forEach(action);
-    }
-
-    public static void parallelForLong(long startInclusive, long endExclusive, LongConsumer action) {
-        if (startInclusive == 0 && endExclusive == 1) {
-            action.accept(0);
-            return;
-        }
-        LongStream.range(startInclusive, endExclusive).parallel().forEach(action);
-    }
-}
-
-record Pair<First, Second>(First first, Second second) {
-}
-
-record GGMLTensorEntry(MemorySegment mappedFile, String name, GGMLType ggmlType, int[] shape,
-                       MemorySegment memorySegment) {
-}
-
-enum GGMLType {
-    F32(Float.BYTES),
-    F16(GGMLType.FLOAT16_BYTES),
-    Q4_0(GGMLType.FLOAT16_BYTES + 16 * Byte.BYTES, 32),
-    Q4_1(2 * GGMLType.FLOAT16_BYTES + 16 * Byte.BYTES, 32),
-    UNSUPPORTED_Q4_2(Integer.MAX_VALUE), // support has been removed
-    UNSUPPORTED_Q4_3(Integer.MAX_VALUE), // support has been removed
-    Q5_0(Integer.MAX_VALUE),
-    Q5_1(Integer.MAX_VALUE),
-    Q8_0(GGMLType.FLOAT16_BYTES + 32 * Byte.BYTES, 32),
-    Q8_1(32 * Byte.BYTES + 2 * Float.BYTES, 32),
-    // k-quantizations
-    Q2_K(Integer.MAX_VALUE),
-    Q3_K(Integer.MAX_VALUE),
-    Q4_K(2 * GGMLType.FLOAT16_BYTES + ((GGMLType.QK_K / 16) / 8 * 6) + GGMLType.QK_K / 2, GGMLType.QK_K),
-    Q5_K(2 * GGMLType.FLOAT16_BYTES + ((GGMLType.QK_K / 16) / 8 * 6) + GGMLType.QK_K / 8 + GGMLType.QK_K / 2, GGMLType.QK_K),
-    Q6_K(GGMLType.QK_K / 2 + GGMLType.QK_K / 4 + GGMLType.QK_K / 16 + GGMLType.FLOAT16_BYTES, GGMLType.QK_K),
-    Q8_K(Integer.MAX_VALUE),
-
-    IQ2_XXS(Integer.MAX_VALUE),
-    IQ2_XS(Integer.MAX_VALUE),
-    IQ3_XXS(Integer.MAX_VALUE),
-    IQ1_S(Integer.MAX_VALUE),
-    IQ4_NL(Integer.MAX_VALUE),
-    IQ3_S(Integer.MAX_VALUE),
-    IQ2_S(Integer.MAX_VALUE),
-    IQ4_XS(Integer.MAX_VALUE),
-
-    I8(Byte.BYTES),
-    I16(Short.BYTES),
-    I32(Integer.BYTES),
-    I64(Long.BYTES),
-    F64(Double.BYTES),
-    IQ1_M(Integer.MAX_VALUE),
-    BF16(GGMLType.BFLOAT16_BYTES),
-    Q4_0_4_4(GGMLType.FLOAT16_BYTES + 16 * Byte.BYTES, 32),
-    Q4_0_4_8(GGMLType.FLOAT16_BYTES + 16 * Byte.BYTES, 32),
-    Q4_0_8_8(GGMLType.FLOAT16_BYTES + 16 * Byte.BYTES, 32),
-    TQ1_0(Integer.MAX_VALUE),
-    TQ2_0(Integer.MAX_VALUE);
-
-    public static final int BFLOAT16_BYTES = 2;
-    public static final int FLOAT16_BYTES = 2;
-
-    private static final GGMLType[] VALUES = values();
-
-    private final int typeSize;
-
-    private final int blockSize;
-
-    public int getTypeSize() {
-        return typeSize;
-    }
-
-    public int getBlockSize() {
-        return blockSize;
-    }
-
-    public static GGMLType fromId(int id) {
-        return VALUES[id];
-    }
-
-    GGMLType(int typeSize) {
-        this(typeSize, 1);
-    }
-
-    public long byteSizeFor(int numberOfElements) {
-        long t = numberOfElements * (long) getTypeSize();
-        assert t % getBlockSize() == 0;
-        return Math.toIntExact(t / getBlockSize());
-    }
-
-    public static final int QK_K = 256; // or 64?
-
-    GGMLType(int typeSize, int blockSize) {
-        assert blockSize > 0;
-        assert typeSize > 0;
-        assert isPowerOf2(blockSize);
-        this.typeSize = typeSize;
-        this.blockSize = blockSize;
-    }
-
-    private static boolean isPowerOf2(int n) {
-        return n > 0 && (n & (n - 1)) == 0;
-    }
-}
-
-/**
- * Over-simplified, shapeless, float tensor.
- * <p>
- * Not a strict tensor, but rather just a sequence of floats, not required to be backed by memory
- * e.g. can represent a sequence of quantized floats.
- */
-abstract class FloatTensor {
-    static final int VECTOR_BIT_SIZE = Integer.getInteger("llama.VectorBitSize", VectorShape.preferredShape().vectorBitSize());
-    static final boolean USE_VECTOR_API = VECTOR_BIT_SIZE != 0;
-
-    // static final ValueLayout.OfFloat JAVA_FLOAT_LE = ValueLayout.JAVA_FLOAT.withOrder(ByteOrder.LITTLE_ENDIAN);
-    // static final ValueLayout.OfShort JAVA_SHORT_LE = ValueLayout.JAVA_SHORT.withOrder(ByteOrder.LITTLE_ENDIAN);
-
-    // The use of Unsafe in this file is a temporary workaround to support native-image.
-    static final Unsafe UNSAFE;
-
-    static {
-        try {
-            Field f = Unsafe.class.getDeclaredField("theUnsafe");
-            f.setAccessible(true);
-            UNSAFE = (Unsafe) f.get(null);
-        } catch (NoSuchFieldException | IllegalAccessException e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-    static short readShort(MemorySegment memorySegment, long offset) {
-        // The MemorySegment.get* methods should be used instead.
-        return UNSAFE.getShort(memorySegment.address() + offset);
-    }
-
-    static byte readByte(MemorySegment memorySegment, long offset) {
-        // The MemorySegment.get* methods should be used instead.
-        return UNSAFE.getByte(memorySegment.address() + offset);
-    }
-
-    // Preferred vector size for the fast multiplication routines.
-    // (Apple Silicon) NEON only supports up-to 128bit vectors.
-    static final VectorSpecies<Float> F_SPECIES;
-    static final VectorSpecies<Integer> I_SPECIES;
-    static final VectorSpecies<Short> S_SPECIES_HALF;
-
-    static {
-        if (USE_VECTOR_API) {
-            F_SPECIES = VectorShape.forBitSize(VECTOR_BIT_SIZE).withLanes(float.class);
-            I_SPECIES = F_SPECIES.withLanes(int.class);
-            S_SPECIES_HALF = VectorShape.forBitSize(F_SPECIES.vectorBitSize() / 2).withLanes(short.class);
-            assert F_SPECIES.length() == S_SPECIES_HALF.length();
-        } else {
-            F_SPECIES = null;
-            I_SPECIES = null;
-            S_SPECIES_HALF = null;
-        }
-    }
-
-    abstract int size();
-
-    abstract float getFloat(int index);
-
-    abstract void setFloat(int index, float value);
-
-    abstract FloatVector getFloatVector(VectorSpecies<Float> species, int offset);
-
-    abstract GGMLType type();
-
-    public static int numberOfElements(int... dimensions) {
-        assert Arrays.stream(dimensions).allMatch(i -> i > 0);
-        return Arrays.stream(dimensions).reduce(Math::multiplyExact).orElseThrow();
-    }
-
-    static float scalarDot(FloatTensor thiz, int thisOffset, FloatTensor that, int thatOffset, int size) {
-        float result = 0f;
-        for (int j = 0; j < size; j++) {
-            result += thiz.getFloat(thisOffset + j) * that.getFloat(thatOffset + j);
-        }
-        return result;
-    }
-
-    float dot(int thisOffset, FloatTensor that, int thatOffset, int size) {
-        return scalarDot(this, thisOffset, that, thatOffset, size);
-    }
-
-    void matmul(FloatTensor that, FloatTensor out, int dim0, int dim1) {
-        Parallel.parallelFor(0, dim0, i -> out.setFloat(i, dot(i * dim1, that, 0, dim1)));
-    }
-
-    void matmul(int context, FloatTensor[] that, FloatTensor[] out, int dim0, int dim1) {
-        if (that.length != out.length) {
-            throw new IllegalArgumentException(String.format("that.len=%d, out.len=%d", that.length, out.length));
-        }
-        Parallel.parallelForLong(0, dim0 * context, ti -> {
-            int idxArr = (int) (ti / dim0);
-            int i = (int) (ti % dim0);
-            out[idxArr].setFloat(i, dot(i * dim1, that[idxArr], 0, dim1)); 
-        });
-    }
-
-    @FunctionalInterface
-    interface AggregateFunction {
-        float apply(float acc, float value);
-    }
-
-    float reduce(int thisOffset, int size, float seed, AggregateFunction reduce) {
-        float result = seed;
-        for (int i = 0; i < size; ++i) {
-            result = reduce.apply(result, getFloat(thisOffset + i));
-        }
-        return result;
-    }
-
-    float sum(int thisOffset, int size) {
-        return reduce(thisOffset, size, 0f, Float::sum);
-    }
-
-    float max(int thisOffset, int size) {
-        return reduce(thisOffset, size, Float.NEGATIVE_INFINITY, Float::max);
-    }
-
-    void copyTo(int thisOffset, FloatTensor that, int thatOffset, int size) {
-        that.mapWithIndexInPlace(thatOffset, size, (value, index) -> this.getFloat(index - thatOffset + thisOffset));
-    }
-
-    int argmax(int thisOffset, int size) {
-        assert size > 0;
-        int maxIndex = thisOffset;
-        float maxValue = this.getFloat(maxIndex);
-        int endIndex = thisOffset + size;
-        for (int i = thisOffset; i < endIndex; ++i) {
-            float f = this.getFloat(i);
-            if (f > maxValue) {
-                maxValue = f;
-                maxIndex = i;
-            }
-        }
-        return maxIndex;
-    }
-
-    int argmax() {
-        return argmax(0, size());
-    }
-
-    @FunctionalInterface
-    interface MapFunction {
-        float apply(float value);
-    }
-
-    @FunctionalInterface
-    interface MapWithIndexFunction {
-        float apply(float value, int index);
-    }
-
-    FloatTensor mapInPlace(int thisOffset, int size, MapFunction mapFunction) {
-        int endIndex = thisOffset + size;
-        for (int i = thisOffset; i < endIndex; ++i) {
-            setFloat(i, mapFunction.apply(getFloat(i)));
-        }
-        return this;
-    }
-
-    FloatTensor mapInPlace(MapFunction mapFunction) {
-        return mapInPlace(0, size(), mapFunction);
-    }
-
-    FloatTensor mapWithIndexInPlace(int thisOffset, int size, FloatTensor.MapWithIndexFunction mapWithIndexFunction) {
-        int endOffset = thisOffset + size;
-        for (int i = thisOffset; i < endOffset; ++i) {
-            setFloat(i, mapWithIndexFunction.apply(getFloat(i), i));
-        }
-        return this;
-    }
-
-    FloatTensor addInPlace(int thisOffset, FloatTensor that, int thatOffset, int size) {
-        return mapWithIndexInPlace(thisOffset, size, (value, index) -> value + that.getFloat(index - thisOffset + thatOffset));
-    }
-
-    FloatTensor addInPlace(FloatTensor that) {
-        return addInPlace(0, that, 0, size());
-    }
-
-    FloatTensor multiplyInPlace(int thisOffset, FloatTensor that, int thatOffset, int size) {
-        return mapWithIndexInPlace(thisOffset, size, (value, index) -> value * that.getFloat(index - thisOffset + thatOffset));
-    }
-
-    FloatTensor multiplyInPlace(FloatTensor that) {
-        return multiplyInPlace(0, that, 0, size());
-    }
-
-    FloatTensor divideInPlace(int thisOffset, int size, float value) {
-        return mapInPlace(thisOffset, size, f -> f / value);
-    }
-
-    FloatTensor fillInPlace(int thisOffset, int size, float value) {
-        return mapInPlace(thisOffset, size, unused -> value);
-    }
-
-    FloatTensor softmaxInPlace(int thisOffset, int size) {
-        // find max value (for numerical stability)
-        float maxVal = max(thisOffset, size);
-        // exp and sum
-        mapInPlace(thisOffset, size, f -> (float) Math.exp(f - maxVal));
-        float sum = sum(thisOffset, size);
-        // normalize
-        return divideInPlace(thisOffset, size, sum);
-    }
-
-    FloatTensor saxpyInPlace(int thisOffset, FloatTensor that, int thatOffset, int size, float a) {
-        // this[thatOffset ... thatOffset + size) = a * that[thatOffset ... thatOffset + size) + this[thisOffset ... thisOffset + size)
-        for (int i = 0; i < size; ++i) {
-            setFloat(thisOffset + i, a * that.getFloat(thatOffset + i) + this.getFloat(thisOffset + i));
-        }
-        return this;
-    }
-}
-
-/**
- * {@link FloatTensor} quantized in the {@link GGMLType#Q4_0} format.
- * <p>
- * This tensor implementation is not compatible with {@link FloatTensor}, but
- * {@link #dot(int, FloatTensor, int, int)} has a vectorized implementation that is used when
- * the second argument implements {@link FloatTensor}.
- */
-final class Q4_0FloatTensor extends FloatTensor {
-
-    final int size;
-    final MemorySegment memorySegment;
-
-    public Q4_0FloatTensor(int size, MemorySegment memorySegment) {
-        this.size = size;
-        this.memorySegment = memorySegment;
-    }
-
-    @Override
-    int size() {
-        return size;
-    }
-
-    @Override
-    public void setFloat(int index, float value) {
-        throw new UnsupportedOperationException("setFloat");
-    }
-
-    @Override
-    FloatVector getFloatVector(VectorSpecies<Float> species, int index) {
-        throw new UnsupportedOperationException("getFloatVector");
-    }
-
-    @Override
-    public GGMLType type() {
-        return GGMLType.Q4_0;
-    }
-
-    @Override
-    public float getFloat(int index) {
-        assert 0 <= index && index < size;
-        int blockIndex = index / GGMLType.Q4_0.getBlockSize();
-        int blockOffset = blockIndex * GGMLType.Q4_0.getTypeSize();
-        float scale = Float.float16ToFloat(readShort(memorySegment, blockOffset));
-        byte quant;
-        int modIndex = index % GGMLType.Q4_0.getBlockSize();
-        if (modIndex < GGMLType.Q4_0.getBlockSize() / 2) {
-            quant = (byte) (readByte(memorySegment, blockOffset + GGMLType.FLOAT16_BYTES + modIndex) & 0x0F);
-        } else {
-            quant = (byte) ((readByte(memorySegment, blockOffset + GGMLType.FLOAT16_BYTES + modIndex - GGMLType.Q4_0.getBlockSize() / 2) >>> 4) & 0x0F);
-        }
-        quant -= 8;
-        return quant * scale;
-    }
-
-    @Override
-    public float dot(int thisOffset, FloatTensor that, int thatOffset, int size) {
-        if (FloatTensor.USE_VECTOR_API) {
-            return vectorDot(this, thisOffset, (ArrayFloatTensor) that, thatOffset, size);
-        } else {
-            return FloatTensor.scalarDot(this, thisOffset, that, thatOffset, size);
-        }
-    }
-
-    private static float vectorDot(Q4_0FloatTensor thiz, int thisOffset, ArrayFloatTensor that, int thatOffset, int size) {
-        float result = 0f;
-        int j = 0;
-
-        // Align thisOffset + j to type().getBlockSize().
-        assert Integer.bitCount(GGMLType.Q4_0.getBlockSize()) == 1 : "power of 2";
-        int alignmentBound = Math.min(size, -thisOffset & (GGMLType.Q4_0.getBlockSize() - 1));
-        if (alignmentBound > 0) {
-            result += FloatTensor.scalarDot(thiz, thisOffset, that, thatOffset, alignmentBound);
-            j += alignmentBound;
-        }
-        assert (thisOffset + j) % GGMLType.Q4_0.getBlockSize() == 0;
-
-        FloatVector val = FloatVector.zero(F_SPECIES);
-        int blockOffset = (thisOffset + j) / GGMLType.Q4_0.getBlockSize() * GGMLType.Q4_0.getTypeSize();
-        int upperBound = size / GGMLType.Q4_0.getBlockSize() * GGMLType.Q4_0.getBlockSize();
-        for (; j < upperBound; j += GGMLType.Q4_0.getBlockSize(), blockOffset += GGMLType.Q4_0.getTypeSize()) {
-            float wScaleValue = Float.float16ToFloat(readShort(thiz.memorySegment, blockOffset));
-            var wScale = FloatVector.broadcast(F_SPECIES, wScaleValue);
-            var wBytes = ByteVector.fromMemorySegment(ByteVector.SPECIES_128, thiz.memorySegment, blockOffset + GGMLType.FLOAT16_BYTES, ByteOrder.LITTLE_ENDIAN);
-            var loBytes = wBytes.and((byte) 0xF).sub((byte) 8);
-            var hiBytes = wBytes.lanewise(VectorOperators.LSHR, 4).sub((byte) 8);
-            switch (F_SPECIES.vectorBitSize()) {
-                case 512 -> {
-                    var sum0 = that.getFloatVector(F_SPECIES, thatOffset + j + 0 * F_SPECIES.length()).mul(loBytes.castShape(F_SPECIES, 0));
-                    var sum2 = that.getFloatVector(F_SPECIES, thatOffset + j + 1 * F_SPECIES.length()).mul(hiBytes.castShape(F_SPECIES, 0));
-                    val = sum0.add(sum2).fma(wScale, val);
-                }
-                case 256 -> {
-                    var sum0 = that.getFloatVector(F_SPECIES, thatOffset + j + 0 * F_SPECIES.length()).mul(loBytes.castShape(F_SPECIES, 0));
-                    var sum1 = that.getFloatVector(F_SPECIES, thatOffset + j + 1 * F_SPECIES.length()).mul(loBytes.castShape(F_SPECIES, 1));
-                    var sum2 = that.getFloatVector(F_SPECIES, thatOffset + j + 2 * F_SPECIES.length()).mul(hiBytes.castShape(F_SPECIES, 0));
-                    var sum3 = that.getFloatVector(F_SPECIES, thatOffset + j + 3 * F_SPECIES.length()).mul(hiBytes.castShape(F_SPECIES, 1));
-                    val = sum0.add(sum1).add(sum2).add(sum3).fma(wScale, val);
-                }
-                case 128 -> {
-                    // This loop cannot be unrolled, why?
-                    for (int i = 0; i < 2; ++i) {
-                        var tmp = i == 0 ? loBytes : hiBytes;
-                        var sum0 = that.getFloatVector(F_SPECIES, thatOffset + j + (i * 4 + 0) * F_SPECIES.length()).mul(tmp.castShape(F_SPECIES, 0));
-                        var sum1 = that.getFloatVector(F_SPECIES, thatOffset + j + (i * 4 + 1) * F_SPECIES.length()).mul(tmp.castShape(F_SPECIES, 1));
-                        var sum2 = that.getFloatVector(F_SPECIES, thatOffset + j + (i * 4 + 2) * F_SPECIES.length()).mul(tmp.castShape(F_SPECIES, 2));
-                        var sum3 = that.getFloatVector(F_SPECIES, thatOffset + j + (i * 4 + 3) * F_SPECIES.length()).mul(tmp.castShape(F_SPECIES, 3));
-                        val = sum0.add(sum1).add(sum2).add(sum3).fma(wScale, val);
-                    }
-                }
-                default -> throw new UnsupportedOperationException(F_SPECIES.toString());
-            }
-        }
-        result += val.reduceLanes(VectorOperators.ADD);
-
-        // Remaining entries.
-        if (j < size) {
-            result += FloatTensor.scalarDot(thiz, thisOffset + j, that, thatOffset + j, size - j);
-        }
-
-        return result;
-    }
-}
-
-final class Q8_0FloatTensor extends FloatTensor {
-
-    final int size;
-    final MemorySegment memorySegment;
-
-    public Q8_0FloatTensor(int size, MemorySegment memorySegment) {
-        this.size = size;
-        this.memorySegment = memorySegment;
-    }
-
-    @Override
-    int size() {
-        return size;
-    }
-
-    @Override
-    public void setFloat(int index, float value) {
-        throw new UnsupportedOperationException("setFloat");
-    }
-
-    @Override
-    FloatVector getFloatVector(VectorSpecies<Float> species, int index) {
-        throw new UnsupportedOperationException("getFloatVector");
-    }
-
-    @Override
-    public GGMLType type() {
-        return GGMLType.Q8_0;
-    }
-
-    @Override
-    public float getFloat(int index) {
-        assert 0 <= index && index < size;
-        int blockIndex = index / GGMLType.Q8_0.getBlockSize();
-        int withinBlockIndex = index % GGMLType.Q8_0.getBlockSize();
-        int blockOffset = blockIndex * GGMLType.Q8_0.getTypeSize();
-        byte quant = readByte(memorySegment, blockOffset + GGMLType.FLOAT16_BYTES + withinBlockIndex);
-        float scale = Float.float16ToFloat(readShort(memorySegment, blockOffset));
-        return quant * scale;
-    }
-
-    public static final ValueLayout.OfShort JAVA_SHORT_LE = ValueLayout.JAVA_SHORT.withOrder(ByteOrder.LITTLE_ENDIAN);
-
-    @Override
-    public float dot(int thisOffset, FloatTensor that, int thatOffset, int size) {
-        if (FloatTensor.USE_VECTOR_API) {
-            return vectorDot(this, thisOffset, (ArrayFloatTensor) that, thatOffset, size);
-        } else {
-            return FloatTensor.scalarDot(this, thisOffset, that, thatOffset, size);
-        }
-    }
-
-    private static float vectorDot(Q8_0FloatTensor thiz, int thisOffset, ArrayFloatTensor that, int thatOffset, int size) {
-        float result = 0f;
-        int j = 0;
-
-        // Align thisOffset + startIndex to type().getBlockSize().
-        assert Integer.bitCount(GGMLType.Q8_0.getBlockSize()) == 1 : "power of 2";
-        int alignmentBound = Math.min(size, -thisOffset & (GGMLType.Q8_0.getBlockSize() - 1));
-        if (alignmentBound > 0) {
-            result += FloatTensor.scalarDot(thiz, thisOffset, that, thatOffset, alignmentBound);
-            j += alignmentBound;
-        }
-        assert (thisOffset + j) % GGMLType.Q8_0.getBlockSize() == 0;
-
-        FloatVector val = FloatVector.zero(F_SPECIES);
-        int blockOffset = (thisOffset + j) / GGMLType.Q8_0.getBlockSize() * GGMLType.Q8_0.getTypeSize();
-        int upperBound = size / GGMLType.Q8_0.getBlockSize() * GGMLType.Q8_0.getBlockSize();
-        for (; j < upperBound; j += GGMLType.Q8_0.getBlockSize(), blockOffset += GGMLType.Q8_0.getTypeSize()) {
-            float wScaleValue = Float.float16ToFloat(readShort(thiz.memorySegment, blockOffset));
-            var wScale = FloatVector.broadcast(F_SPECIES, wScaleValue);
-            switch (F_SPECIES.vectorBitSize()) {
-                case 512 -> {
-                    var wBytes = ByteVector.fromMemorySegment(ByteVector.SPECIES_256, thiz.memorySegment, blockOffset + GGMLType.FLOAT16_BYTES, ByteOrder.LITTLE_ENDIAN);
-                    var sum0 = that.getFloatVector(F_SPECIES, thatOffset + j + 0 * F_SPECIES.length()).mul(wBytes.castShape(F_SPECIES, 0));
-                    var sum1 = that.getFloatVector(F_SPECIES, thatOffset + j + 1 * F_SPECIES.length()).mul(wBytes.castShape(F_SPECIES, 1));
-                    val = sum0.add(sum1).fma(wScale, val);
-                }
-                case 256 -> {
-                    var wBytes = ByteVector.fromMemorySegment(ByteVector.SPECIES_256, thiz.memorySegment, blockOffset + GGMLType.FLOAT16_BYTES, ByteOrder.LITTLE_ENDIAN);
-                    var sum0 = that.getFloatVector(F_SPECIES, thatOffset + j + 0 * F_SPECIES.length()).mul(wBytes.castShape(F_SPECIES, 0));
-                    var sum1 = that.getFloatVector(F_SPECIES, thatOffset + j + 1 * F_SPECIES.length()).mul(wBytes.castShape(F_SPECIES, 1));
-                    var sum2 = that.getFloatVector(F_SPECIES, thatOffset + j + 2 * F_SPECIES.length()).mul(wBytes.castShape(F_SPECIES, 2));
-                    var sum3 = that.getFloatVector(F_SPECIES, thatOffset + j + 3 * F_SPECIES.length()).mul(wBytes.castShape(F_SPECIES, 3));
-                    val = sum0.add(sum1).add(sum2).add(sum3).fma(wScale, val);
-                }
-                case 128 -> {
-                    // This loop cannot be unrolled, why?
-                    for (int i = 0; i < 2; ++i) {
-                        var wBytes = ByteVector.fromMemorySegment(ByteVector.SPECIES_128, thiz.memorySegment, blockOffset + GGMLType.FLOAT16_BYTES + i * ByteVector.SPECIES_128.vectorByteSize(), ByteOrder.LITTLE_ENDIAN);
-                        var sum0 = that.getFloatVector(F_SPECIES, thatOffset + j + i * 16 + 0 * F_SPECIES.length()).mul(wBytes.castShape(F_SPECIES, 0));
-                        var sum1 = that.getFloatVector(F_SPECIES, thatOffset + j + i * 16 + 1 * F_SPECIES.length()).mul(wBytes.castShape(F_SPECIES, 1));
-                        var sum2 = that.getFloatVector(F_SPECIES, thatOffset + j + i * 16 + 2 * F_SPECIES.length()).mul(wBytes.castShape(F_SPECIES, 2));
-                        var sum3 = that.getFloatVector(F_SPECIES, thatOffset + j + i * 16 + 3 * F_SPECIES.length()).mul(wBytes.castShape(F_SPECIES, 3));
-                        val = sum0.add(sum1).add(sum2).add(sum3).fma(wScale, val);
-                    }
-                }
-                default -> throw new UnsupportedOperationException(F_SPECIES.toString());
-            }
-        }
-        result += val.reduceLanes(VectorOperators.ADD);
-
-        // Remaining entries.
-        if (j < size) {
-            result += FloatTensor.scalarDot(thiz, thisOffset + j, that, thatOffset + j, size - j);
-        }
-
-        return result;
-    }
-}
-
-final class BF16FloatTensor extends FloatTensor {
-
-    final int size;
-    final MemorySegment memorySegment;
-
-    public BF16FloatTensor(int size, MemorySegment memorySegment) {
-        this.size = size;
-        this.memorySegment = memorySegment;
-    }
-
-    @Override
-    int size() {
-        return size;
-    }
-
-    @Override
-    public void setFloat(int index, float value) {
-        throw new UnsupportedOperationException("setFloat");
-    }
-
-    @Override
-    FloatVector getFloatVector(VectorSpecies<Float> species, int index) {
-        throw new UnsupportedOperationException("getFloatVector");
-    }
-
-    @Override
-    public GGMLType type() {
-        return GGMLType.BF16;
-    }
-
-    @Override
-    public float getFloat(int index) {
-        assert 0 <= index && index < size;
-        return bfloat16ToFloat(readShort(memorySegment, index * GGMLType.BFLOAT16_BYTES));
-    }
-
-    private float bfloat16ToFloat(short bfloat16) {
-        return Float.intBitsToFloat(bfloat16 << 16);
-    }
-
-    @Override
-    public float dot(int thisOffset, FloatTensor that, int thatOffset, int size) {
-        if (FloatTensor.USE_VECTOR_API) {
-            return vectorDot(this, thisOffset, (ArrayFloatTensor) that, thatOffset, size);
-        } else {
-            return FloatTensor.scalarDot(this, thisOffset, that, thatOffset, size);
-        }
-    }
-
-    private static float vectorDot(BF16FloatTensor thiz, int thisOffset, ArrayFloatTensor that, int thatOffset, int size) {
-        assert S_SPECIES_HALF.length() == F_SPECIES.length();
-        FloatVector val = FloatVector.zero(F_SPECIES);
-        int upperBound = F_SPECIES.loopBound(size);
-        for (int i = 0; i < upperBound; i += F_SPECIES.length()) {
-            FloatVector thatVector = that.getFloatVector(F_SPECIES, thatOffset + i);
-            ShortVector bfloat16 = ShortVector.fromMemorySegment(S_SPECIES_HALF, thiz.memorySegment, (thisOffset + i) * (long) GGMLType.BFLOAT16_BYTES, ByteOrder.LITTLE_ENDIAN);
-            // BFloat16 to Float32 Conversion:
-            //
-            // ┌─[15]─┬─[14]───····───[7]─┬─[6]────····────[0]─┐
-            // │ Sign │ Exponent (8 bits) │ Mantissa (7 bits)  │ BFloat16 Layout (16 bits)
-            // └──────┴───────────────────┴────────────────────┘
-            //    │             │                    │
-            //    ▼             ▼                    ▼
-            // ┌─[31]─┬─[30]───···───[23]─┬─[22]────···────[0]─┐
-            // │ Sign │ Exponent (8 bits) │ Mantissa (23 bits) │ Float32 Layout (32 bits)
-            // └──────┴───────────────────┴────────────────────┘
-            FloatVector thizVector = bfloat16
-                    .castShape(I_SPECIES, 0) // (int) vi
-                    .lanewise(VectorOperators.LSHL, 16) // vi <<= 16
-                    .reinterpretAsFloats(); // Float.intBitsToFloat(vi)
-            val = thizVector.fma(thatVector, val);
-        }
-        float result = val.reduceLanes(VectorOperators.ADD);
-        // Remaining entries.
-        if (upperBound < size) {
-            result += scalarDot(thiz, thisOffset + upperBound, that, thatOffset + upperBound, size - upperBound);
-        }
-
-        return result;
-    }
-}
-
-final class F16FloatTensor extends FloatTensor {
-
-    final int size;
-    final MemorySegment memorySegment;
-
-    public F16FloatTensor(int size, MemorySegment memorySegment) {
-        this.size = size;
-        this.memorySegment = memorySegment;
-    }
-
-    @Override
-    int size() {
-        return size;
-    }
-
-    @Override
-    public void setFloat(int index, float value) {
-        throw new UnsupportedOperationException("setFloat");
-    }
-
-    @Override
-    FloatVector getFloatVector(VectorSpecies<Float> species, int index) {
-        throw new UnsupportedOperationException("getFloatVector");
-    }
-
-    @Override
-    public GGMLType type() {
-        return GGMLType.F16;
-    }
-
-    @Override
-    public float getFloat(int index) {
-        assert 0 <= index && index < size;
-        return Float.float16ToFloat(readShort(memorySegment, index * GGMLType.FLOAT16_BYTES));
-    }
-
-    @Override
-    public float dot(int thisOffset, FloatTensor that, int thatOffset, int size) {
-        if (FloatTensor.USE_VECTOR_API) {
-            return vectorDot(this, thisOffset, (ArrayFloatTensor) that, thatOffset, size);
-        } else {
-            return FloatTensor.scalarDot(this, thisOffset, that, thatOffset, size);
-        }
-    }
-
-    private static float vectorDot(F16FloatTensor thiz, int thisOffset, ArrayFloatTensor that, int thatOffset, int size) {
-        assert S_SPECIES_HALF.length() == F_SPECIES.length();
-        FloatVector val = FloatVector.zero(F_SPECIES);
-        int upperBound = F_SPECIES.loopBound(size);
-        for (int i = 0; i < upperBound; i += F_SPECIES.length()) {
-            FloatVector thatVector = that.getFloatVector(F_SPECIES, thatOffset + i);
-            ShortVector bits16 = ShortVector.fromMemorySegment(S_SPECIES_HALF, thiz.memorySegment, (thisOffset + i) * (long) GGMLType.FLOAT16_BYTES, ByteOrder.LITTLE_ENDIAN);
-
-            var bits32 = bits16.castShape(I_SPECIES, 0).reinterpretAsInts(); // (int) bits16
-            // Does not support infinities nor NaNs, preserves sign, emulate DAZ (denormals-are-zero).
-            // Expects well-formed float16 values only (e.g. model weights).
-            // Fast Float16 to Float32 Conversion:
-            //
-            // ┌─[15]─┬─[14]───···───[10]─┬─[9]────····────[0]─┐
-            // │ Sign │ Exponent (5 bits) │ Mantissa (10 bits) │ Float16 Layout (16 bits)
-            // └──────┴───────────────────┴────────────────────┘
-            //    │             │                    │
-            //    ▼             ▼                    ▼
-            // ┌─[31]─┬─[30]───···───[23]─┬─[22]────···────[0]─┐
-            // │ Sign │ Exponent (8 bits) │ Mantissa (23 bits) │ Float32 Layout (32 bits)
-            // └──────┴───────────────────┴────────────────────┘
-            //
-            // Shifts and adjustments:
-            // - Sign:       float16[15] -> float32[31] (shift 16 bits up)
-            // - Exponent:   float16[10-14] -> float32[23-30] (+ bias adjustment)
-            // - Mantissa:   float16[0-9] -> float32[13-22] (shift 13 bits up)
-            //
-            // exp = bits32 & 0x7C00
-            // zeroExponentMask = exp == 0 ? 0 : ~0
-            var zeroExponentMask = bits32.and(0x7C00).neg().lanewise(VectorOperators.ASHR, 31); // = (-exp) >> 31
-            bits32 = bits32.and(0x8000).lanewise(VectorOperators.LSHL, 16) // sign
-                    .or(
-                            // exponent and mantissa combined
-                            bits32.and(0x7FFF).add(0x1C000).lanewise(VectorOperators.LSHL, 13)
-                                    .and(zeroExponentMask) // -0, +0 and DAZ (denormals-are-zero)
-
-                    );
-
-            FloatVector thizVector = bits32.reinterpretAsFloats(); // Float.intBitsToFloat(vi)
-            val = thizVector.fma(thatVector, val);
-        }
-        float result = val.reduceLanes(VectorOperators.ADD);
-        // Remaining entries.
-        if (upperBound < size) {
-            result += scalarDot(thiz, thisOffset + upperBound, that, thatOffset + upperBound, size - upperBound);
-        }
-
-        return result;
-    }
-}
-
-final class ArrayFloatTensor extends FloatTensor {
-
-    final float[] values;
-
-    ArrayFloatTensor(float[] values) {
-        this.values = values;
-    }
-
-    public static FloatTensor allocate(int... dims) {
-        int numberOfElements = FloatTensor.numberOfElements(dims);
-        return new ArrayFloatTensor(new float[numberOfElements]);
-    }
-
-    @Override
-    public int size() {
-        return values.length;
-    }
-
-    @Override
-    public float getFloat(int index) {
-        return values[index];
-    }
-
-    @Override
-    public void setFloat(int index, float value) {
-        values[index] = value;
-    }
-
-    @Override
-    public GGMLType type() {
-        return GGMLType.F32;
-    }
-
-    @Override
-    public FloatTensor fillInPlace(int thisOffset, int size, float value) {
-        Arrays.fill(values, thisOffset, thisOffset + size, value);
-        return this;
-    }
-
-    @Override
-    public FloatVector getFloatVector(VectorSpecies<Float> species, int index) {
-        if (!USE_VECTOR_API) {
-            throw new UnsupportedOperationException();
-        }
-        return FloatVector.fromArray(species, values, index);
-    }
-}
-
-final class RoPE {
-    public static Pair<float[], float[]> precomputeFreqsCis(int contextLength, int headSize, double theta,
-                                                            boolean ropeScaling, float scaleFactor, float loFreqFactor, float hiFreqFactor, float oldContextLength) {
-        assert headSize % 2 == 0;
-        float[] cr = new float[contextLength * (headSize / 2)];
-        float[] ci = new float[contextLength * (headSize / 2)];
-        int n = 0;
-        for (int pos = 0; pos < contextLength; ++pos) {
-            for (int i = 0; i < headSize; i += 2) {
-                float freq = (float) (1.0 / Math.pow(theta, i / (double) headSize));
-                if (ropeScaling) {
-                    // Llama 3.1 scaling
-                    float loFreqWavelen = oldContextLength / loFreqFactor;
-                    float hiFreqWavelen = oldContextLength / hiFreqFactor;
-                    float wavelen = (float) (2.0 * Math.PI / freq);
-                    if (wavelen < hiFreqWavelen) {
-                        freq = freq;
-                    } else if (wavelen > loFreqWavelen) {
-                        freq = freq / scaleFactor;
-                    } else {
-                        float smooth = (oldContextLength / wavelen - loFreqFactor) / (hiFreqFactor - loFreqFactor);
-                        freq = (1.0f - smooth) * freq / scaleFactor + smooth * freq;
-                    }
-                }
-                float val = pos * freq;
-                cr[n] = (float) Math.cos(val);
-                ci[n] = (float) Math.sin(val);
-                n++;
-            }
-        }
-        assert contextLength * (headSize / 2) == n;
-        return new Pair<>(cr, ci);
-    }
-}
-
-record Vocabulary(String[] tokens, float[] scores, Map<String, Integer> tokenToIndex) {
-    public Vocabulary(String[] vocabulary, float[] scores) {
-        this(vocabulary, scores,
-                IntStream.range(0, vocabulary.length)
-                        .boxed()
-                        .collect(Collectors.toMap(i -> vocabulary[i], i -> i))
-        );
-    }
-
-    public String get(int tokenIndex) {
-        return tokens[tokenIndex];
-    }
-
-    public OptionalInt getIndex(String token) {
-        Integer value = tokenToIndex.get(token);
-        return value != null ? OptionalInt.of(value) : OptionalInt.empty();
-    }
-
-    public int size() {
-        return tokens.length;
-    }
-}
-
-@FunctionalInterface
-interface Sampler {
-    int sampleToken(FloatTensor logits);
-
-    Sampler ARGMAX = FloatTensor::argmax;
-}
-
-record CategoricalSampler(RandomGenerator rng) implements Sampler {
-
-    @Override
-    public int sampleToken(FloatTensor logits) {
-        // sample index from probabilities (they must sum to 1!)
-        float random0to1 = rng.nextFloat(1f);
-        float cdf = 0.0f;
-        for (int i = 0; i < logits.size(); i++) {
-            cdf += logits.getFloat(i);
-            if (random0to1 < cdf) {
-                return i;
-            }
-        }
-        return logits.size() - 1; // in case of rounding errors
-    }
-}
-
-final class ToppSampler implements Sampler {
-
-    final int[] indices;
-    final float topp;
-    final RandomGenerator rng;
-
-    public ToppSampler(int maxNumberOfElements, float topp, RandomGenerator rng) {
-        this.indices = new int[maxNumberOfElements];
-        this.topp = topp;
-        this.rng = rng;
-    }
-
-    static void swap(int[] array, int from, int to) {
-        int tmp = array[from];
-        array[from] = array[to];
-        array[to] = tmp;
-    }
-
-    static void siftDown(int[] array, int from, int n, Comparator<Integer> comparator) {
-        int prev = from, next;
-        while ((next = 2 * prev + 1) < n) {
-            int r = 2 * prev + 2;
-            if (r < n && comparator.compare(array[r], array[next]) < 0) {
-                next = r;
-            }
-            if (comparator.compare(array[next], array[prev]) < 0) {
-                swap(array, prev, next);
-                prev = next;
-            } else {
-                break;
-            }
-        }
-    }
-
-    @Override
-    public int sampleToken(FloatTensor logits) {
-        // top-p sampling (or "nucleus sampling") samples from the smallest set of
-        // tokens that exceed probability topp. This way we never sample tokens that
-        // have very low probabilities and are less likely to go "off the rails".
-        Comparator<Integer> comparator = Comparator.comparingDouble(logits::getFloat).reversed();
-
-        int n = logits.size();
-        int head = 0;
-        int tail = n - 1;
-        // values smaller than (1 - topp) / (n - 1) cannot be part of the result
-        // so for efficiency we crop these out as candidates before sorting
-        float cutoff = (1.0f - topp) / (n - 1);
-        for (int i = 0; i < indices.length; i++) {
-            if (logits.getFloat(i) >= cutoff) {
-                indices[head++] = i;
-            } else {
-                indices[tail--] = i;
-            }
-        }
-
-        int n0 = head;
-        // build heap O(n0)
-        for (int i = n0 / 2 - 1; i >= 0; --i) {
-            siftDown(indices, i, n0, comparator);
-        }
-
-        // truncate the list where cumulative probability of the largest k elements exceeds topp
-        // O(k lg n0)
-        float cumulativeProb = 0.0f;
-        int lastIndex = 0;
-        for (int i = n0 - 1; i >= 0; i--) {
-            swap(indices, 0, i);
-            cumulativeProb += logits.getFloat(indices[i]);
-            if (cumulativeProb > topp) {
-                lastIndex = i;
-                break; // we've exceeded topp by including lastIndex
-            }
-            siftDown(indices, 0, i - 1, comparator);
-        }
-
-        // sample from the truncated list
-        float r = rng.nextFloat(1f) * cumulativeProb;
-        float cdf = 0.0f;
-        for (int i = n0 - 1; i >= lastIndex; i--) {
-            cdf += logits.getFloat(indices[i]);
-            if (r < cdf) {
-                return indices[i];
-            }
-        }
-
-        return indices[lastIndex]; // in case of rounding errors
-    }
-}
-
-/**
- * Utility tailored for Llama 3 instruct prompt format.
- */
-class ChatFormat {
-
-    final Tokenizer tokenizer;
-    final int beginOfText;
-    final int endHeader;
-    final int startHeader;
-    final int endOfTurn;
-    final int endOfText;
-    final int endOfMessage;
-    final Set<Integer> stopTokens;
-
-    public ChatFormat(Tokenizer tokenizer) {
-        this.tokenizer = tokenizer;
-        Map<String, Integer> specialTokens = this.tokenizer.getSpecialTokens();
-        this.beginOfText = specialTokens.get("<|begin_of_text|>");
-        this.startHeader = specialTokens.get("<|start_header_id|>");
-        this.endHeader = specialTokens.get("<|end_header_id|>");
-        this.endOfTurn = specialTokens.get("<|eot_id|>");
-        this.endOfText = specialTokens.get("<|end_of_text|>");
-        this.endOfMessage = specialTokens.getOrDefault("<|eom_id|>", -1); // only in 3.1
-        this.stopTokens = Set.of(endOfText, endOfTurn);
-    }
-
-    public Tokenizer getTokenizer() {
-        return tokenizer;
-    }
-
-    public Set<Integer> getStopTokens() {
-        return stopTokens;
-    }
-
-    public List<Integer> encodeHeader(ChatFormat.Message message) {
-        List<Integer> tokens = new ArrayList<>();
-        tokens.add(startHeader);
-        tokens.addAll(this.tokenizer.encodeAsList(message.role().name()));
-        tokens.add(endHeader);
-        tokens.addAll(this.tokenizer.encodeAsList("\n"));
-        return tokens;
-    }
-
-    public List<Integer> encodeMessage(ChatFormat.Message message) {
-        List<Integer> tokens = this.encodeHeader(message);
-        tokens.addAll(this.tokenizer.encodeAsList(message.content().strip()));
-        tokens.add(endOfTurn);
-        return tokens;
-    }
-
-    public List<Integer> encodeDialogPrompt(boolean appendAssistantTurn, List<ChatFormat.Message> dialog) {
-        List<Integer> tokens = new ArrayList<>();
-        tokens.add(beginOfText);
-        for (ChatFormat.Message message : dialog) {
-            tokens.addAll(this.encodeMessage(message));
-        }
-        if (appendAssistantTurn) {
-            // Add the start of an assistant message for the model to complete.
-            tokens.addAll(this.encodeHeader(new ChatFormat.Message(ChatFormat.Role.ASSISTANT, "")));
-        }
-        return tokens;
-    }
-
-    public record Message(ChatFormat.Role role, String content) {
-    }
-
-    public record Role(String name) {
-        public static ChatFormat.Role SYSTEM = new ChatFormat.Role("system");
-        public static ChatFormat.Role USER = new ChatFormat.Role("user");
-        public static ChatFormat.Role ASSISTANT = new ChatFormat.Role("assistant");
-
-        @Override
-        public String toString() {
-            return name;
-        }
-    }
-}
-
-/**
- * Support for AOT preloading of GGUF metadata with GraalVM's Native Image.
- *
- * <p>
- * To preload a model at build time, pass {@code -Dllama.PreloadGGUF=/path/to/model.gguf}
- * to the native-image builder command. At runtime, the preloaded model will be used
- * iff the specified and preloaded file names (base name) match.
- */
-final class AOT {
-    record PartialModel(String modelFileName, Llama model, long tensorDataOffset, Map<String, GGUF.GGUFTensorInfo> tensorInfos) {}
-
-    private static final PartialModel PRELOADED_GGUF = preLoadGGUF(System.getProperty("llama.PreloadGGUF"));
-
-    private static PartialModel preLoadGGUF(String modelPath) {
-        if (modelPath == null || modelPath.isEmpty()) {
-            return null;
-        }
-        try {
-            Path path = Path.of(modelPath);
-            if (!Files.exists(path) || !Files.isRegularFile(path)) {
-                throw new IllegalArgumentException("Cannot pre-load model: " + path);
-            }
-            GGUF gguf = GGUF.loadModel(path);
-            try (FileChannel fileChannel = FileChannel.open(path, StandardOpenOption.READ)) {
-                return new PartialModel(
-                        path.getFileName().toString(),
-                        ModelLoader.loadModel(fileChannel, gguf, Options.DEFAULT_MAX_TOKENS, false),
-                        gguf.getTensorDataOffset(),
-                        gguf.getTensorInfos()
-                );
-            }
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-    /**
-     * Tries to reuse a compatible AOT preloaded model.
-     * The file name (base name) must match with the preloaded file name.
-     * No checksum/hash is checked for performance reasons.
-     */
-    public static Llama tryUsePreLoaded(Path modelPath, int contextLength) throws IOException {
-        AOT.PartialModel preLoaded = AOT.PRELOADED_GGUF;
-        if (preLoaded == null) {
-            return null; // no pre-loaded model stored
-        }
-        String optionsModel = modelPath.getFileName().toString();
-        String preLoadedModel = preLoaded.modelFileName();
-        if (!Objects.equals(optionsModel, preLoadedModel)) {
-            // Preloaded and specified model file names didn't match.
-            return null;
-        }
-        Llama baseModel = preLoaded.model();
-        try (var timer = Timer.log("Load tensors from pre-loaded model");
-             var fileChannel = FileChannel.open(modelPath, StandardOpenOption.READ)) {
-            // Load only the tensors (mmap slices).
-            Map<String, GGMLTensorEntry> tensorEntries = GGUF.loadTensors(fileChannel, preLoaded.tensorDataOffset(), preLoaded.tensorInfos());
-            Llama.Weights weights = ModelLoader.loadWeights(tensorEntries, baseModel.configuration());
-            return new Llama(baseModel.configuration().withContextLength(contextLength), baseModel.tokenizer(), weights);
-        }
-    }
-}
