+++ Pair.java
@@ -1,0 +1,5 @@
+package com.llama4j;
+
+public record Pair<First, Second>(First first, Second second) {
+}
+
