+++ Vocabulary.java
@@ -1,0 +1,30 @@
+package com.llama4j;
+
+import java.util.Map;
+import java.util.OptionalInt;
+import java.util.stream.Collectors;
+import java.util.stream.IntStream;
+
+record Vocabulary(String[] tokens, float[] scores, Map<String, Integer> tokenToIndex) {
+    public Vocabulary(String[] vocabulary, float[] scores) {
+        this(vocabulary, scores,
+                IntStream.range(0, vocabulary.length)
+                        .boxed()
+                        .collect(Collectors.toMap(i -> vocabulary[i], i -> i))
+        );
+    }
+
+    public String get(int tokenIndex) {
+        return tokens[tokenIndex];
+    }
+
+    public OptionalInt getIndex(String token) {
+        Integer value = tokenToIndex.get(token);
+        return value != null ? OptionalInt.of(value) : OptionalInt.empty();
+    }
+
+    public int size() {
+        return tokens.length;
+    }
+}
+
